Ensure plus/minus uses long not int
git-svn-id: https://threeten.svn.sourceforge.net/svnroot/threeten/trunk/threeten@1304 291d795c-afe8-5c46-8ee5-bf9dd72e1864

+++ src/main/java/javax/time/calendar/LocalTime.java
@@ -1007,7 +1007,7 @@ public final class LocalTime
      * @param daysOverflow  the number of days to store
      * @return the days-overflow, never null
      */
-    public Overflow toOverflow(int daysOverflow) {
+    public Overflow toOverflow(long daysOverflow) {
         return new Overflow(this, daysOverflow);
     }
